move search bar, reduce functionality, move recipes to separate page
