+++ .gitignore
@@ -37,3 +37,4 @@ dist/
 functions
 
 src/functions/index.html
+report.*.json
