+++ src/frontend/App.tsx
@@ -1,16 +1,16 @@
 import React, { useEffect, useState } from 'react';
 import {
   Switch,
-  Route,
-  Redirect
+  Route
 } from 'react-router-dom';
 import AboutPage from './components/pages/about';
-import SearchPage from './components/pages/search';
 import firebase from 'firebase';
 import { AppState } from './lib/common.interface';
 import './index.css';
 import { UserContext } from './lib/auth';
 import ResultsPage from './components/pages/results';
+import LandingPage from './components/pages/landing';
+import { SearchBar } from './components/search-bar';
 
 function App(props: AppState) {
   const [ user, setUser ] = useState<firebase.User | null>(null);
@@ -23,8 +23,9 @@ function App(props: AppState) {
 
   return (
     <UserContext.Provider value={user}>
+      <SearchBar />
       <Switch>
-        <Route exact path="/" component={SearchPage} />
+        <Route exact path="/" component={LandingPage} />
         <Route exact path="/about" component={AboutPage} />
         <Route exact path="/recipe">
           <ResultsPage data={props.initialState.recipe}/>
