--- src/frontend/components/pages/recipe.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export function RecipePage() {
-  return (
-    <div>
-
-    </div>
-  )
-}
