+++ webpack/templates/index.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
+    <title>Shorter Recipes</title>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#303030" />
@@ -8,6 +9,7 @@
       name="description"
       content="A simple application that can extract recipe information and display it in a straight-forward manner."
     />
+    <link rel="canonical" href="https://shorter.recipes/" />
     <link
       href="https://fonts.googleapis.com/css2?family=Roboto+Slab&display=swap"
       rel="stylesheet"
@@ -18,8 +20,10 @@
       integrity="sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk"
       crossorigin="anonymous"
     />
-    <title>Shorter Recipes</title>
     <script>{{initial_state}}</script>
+    <meta property="og:title" content="Shorter Recipes" />
+    <meta property="og:description" content="Come to cook, not read. Use shorter recipes to improve cook time." />
+    <meta property="og:site_name" content="shorter.recipes" />
   </head>
   <body>
     <noscript>
