+++ webpack/webpack.client.config.js
@@ -50,6 +50,7 @@ module.exports = {
         exclude: /node_modules/,
         options: {
           configFile: 'tsconfig.json',
+          transpileOnly: true
         },
       },
       {
