+++ webpack/webpack.server.config.js
@@ -31,6 +31,7 @@ module.exports = {
         loader: 'ts-loader',
         options: {
           configFile: 'tsconfig.json',
+          transpileOnly: true
         },
       },
       {
